Extract validators to separate file
