+++ src/main/scala/info/fingo/spata/schema/CSVSchema.scala
@@ -5,7 +5,6 @@
  */
 package info.fingo.spata.schema
 
-import scala.util.matching.Regex
 import cats.data.Validated
 import cats.data.Validated.{Invalid, Valid}
 import cats.effect.Sync
@@ -20,6 +19,37 @@ class TypedRecord[+A](val data: A, val lineNum: Int, val rowNum: Int)
 
 object TypedRecord {
   def apply[A](data: A, lineNum: Int, rowNum: Int): TypedRecord[A] = new TypedRecord[A](data, lineNum, rowNum)
+}
+
+class CSVSchema[L <: HList: SchemaEnforcer](columns: L) {
+
+  def validate[F[_]: Sync: Logger](implicit enforcer: SchemaEnforcer[L]): Pipe[F, Record, enforcer.Out] =
+    (in: Stream[F, Record]) => in.map(r => validateRecord(r)(enforcer))
+
+  private def validateRecord(r: Record)(implicit enforcer: SchemaEnforcer[L]): enforcer.Out = enforcer(columns, r)
+}
+
+object CSVSchema {
+  def apply[L <: HList: SchemaEnforcer](columns: L) = new CSVSchema(columns)
+}
+
+class Column[K <: StrSng, V: StringParser](val name: K, val validators: Seq[Validator[V]]) {
+  def validate(record: Record): Validated[FieldFlaw, FieldType[K, V]] = {
+    val typeValidated =
+      Validated.fromEither(record.get(name)).leftMap(e => FieldFlaw(name, NotParsed(e.messageCode, e)))
+    val fullyValidated = typeValidated.andThen { v =>
+      validators
+        .map(_.validate(v).leftMap(FieldFlaw(name, _)))
+        .foldLeft(typeValidated)((prev, curr) => prev.andThen(_ => curr))
+    }
+    fullyValidated.map(field[K](_))
+  }
+}
+
+object Column {
+  def apply[A: StringParser](name: String): Column[name.type, A] = new Column(name, Nil)
+  def apply[A: StringParser](name: String, validators: Seq[Validator[A]]): Column[name.type, A] =
+    new Column(name, validators)
 }
 
 trait SchemaEnforcer[L <: HList] extends DepFn2[L, Record] {
@@ -62,75 +92,4 @@ object SchemaEnforcer {
 
   def enforce[L <: HList](columns: L, record: Record)(implicit enforcer: SchemaEnforcer[L]): enforcer.Out =
     enforcer(columns, record)
-}
-
-class CSVSchema[L <: HList: SchemaEnforcer](columns: L) {
-
-  def validate[F[_]: Sync: Logger](implicit enforcer: SchemaEnforcer[L]): Pipe[F, Record, enforcer.Out] =
-    (in: Stream[F, Record]) => in.map(r => validateRecord(r)(enforcer))
-
-  private def validateRecord(r: Record)(implicit enforcer: SchemaEnforcer[L]): enforcer.Out = enforcer(columns, r)
-}
-
-object CSVSchema {
-  def apply[L <: HList: SchemaEnforcer](columns: L) = new CSVSchema(columns)
-}
-
-class Column[K <: StrSng, V: StringParser](val name: K, val validators: Seq[Validator[V]]) {
-  def validate(record: Record): Validated[FieldFlaw, FieldType[K, V]] = {
-    val typeValidated =
-      Validated.fromEither(record.get(name)).leftMap(e => FieldFlaw(name, NotParsed(e.messageCode, e)))
-    val fullyValidated = typeValidated.andThen { v =>
-      validators
-        .map(_.validate(v).leftMap(FieldFlaw(name, _)))
-        .foldLeft(typeValidated)((prev, curr) => prev.andThen(_ => curr))
-    }
-    fullyValidated.map(field[K](_))
-  }
-}
-
-object Column {
-  def apply[A: StringParser](name: String): Column[name.type, A] = new Column(name, Nil)
-  def apply[A: StringParser](name: String, validators: Seq[Validator[A]]): Column[name.type, A] =
-    new Column(name, validators)
-}
-
-trait Validator[A] {
-  def validate(value: A): Validated[ValidationError, A]
-}
-
-class RegexValidator(regex: Regex) extends Validator[String] {
-  def validate(value: String): Validated[ValidationError, String] =
-    Validated.valid(value).ensure[ValidationError](NotRegexConform)(regex.matches)
-}
-object RegexValidator {
-  def apply(regex: Regex): RegexValidator = new RegexValidator(regex)
-  def apply(regex: String): RegexValidator = new RegexValidator(new Regex(regex))
-}
-
-class MinValidator[A: Ordering](min: A) extends Validator[A] {
-  def validate(value: A): Validated[ValidationError, A] =
-    Validated.valid(value).ensure[ValidationError](ValueToSmall)(implicitly[Ordering[A]].lteq(min, _))
-}
-object MinValidator {
-  def apply[A: Ordering](min: A): MinValidator[A] = new MinValidator(min)
-}
-
-class MaxValidator[A: Ordering](max: A) extends Validator[A] {
-  def validate(value: A): Validated[ValidationError, A] =
-    Validated.valid(value).ensure[ValidationError](ValueToLarge)(implicitly[Ordering[A]].gteq(max, _))
-}
-object MaxValidator {
-  def apply[A: Ordering](max: A): MaxValidator[A] = new MaxValidator(max)
-}
-
-class MinMaxValidator[A: Ordering](min: A, max: A) extends Validator[A] {
-  def validate(value: A): Validated[ValidationError, A] = {
-    val minV = MinValidator[A](min)
-    val maxV = MaxValidator[A](max)
-    minV.validate(value).andThen(_ => maxV.validate(value))
-  }
-}
-object MinMaxValidator {
-  def apply[A: Numeric](min: A, max: A): MinMaxValidator[A] = new MinMaxValidator(min, max)
 }
